# Conflicts:
#	Assets/Mods/CrystalMode/Prefabs/Flyer.prefab
#	Assets/Mods/CrystalMode/Prefabs/Runner.prefab
#	Assets/Scenes/CrystalMode.unity

+++ Assets/Mods/MainMode/Scripts/UI/GeneralElelement/Timer/TimerCountRecord.cs
@@ -33,7 +33,7 @@ namespace MainMode
         private void Update()
         {
             _progress += Time.deltaTime;
-            _timer.Output(_progress);
+            _timer.Output((int)_progress);
         }
 
     }
